:truck: move files :fire: remove site-kit folder
