--- site-kit/README.md
@@ -1,3 +1,0 @@
-# site-kit
-
-This is a collection of styles, components and icons used in common by *.svelte.dev websites. It is *not* intended as a general purpose component kit.
