--- site-kit/index.mjs
@@ -1,1 +1,0 @@
-export { default as Docs } from './components/Docs.svelte';
