+++ src/client.js
@@ -1,5 +1,5 @@
 import * as sapper from '@sapper/app';
-import "../site-kit/base.css"
+import "../static/base.css"
 import './utils/i18n'
 
 sapper.start({
