+++ src/components/Docs.svelte
@@ -1,7 +1,7 @@
 <script>
     import {onMount} from 'svelte';
     import GuideContents from './GuideContents.svelte';
-    import {getFragment} from '../../src/utils/navigation';
+    import {getFragment} from '../utils/navigation';
 
     import {_, locale, locales} from 'svelte-i18n';
